melhorado as validaçoes do backend

+++ packages/backend/src/app/controllers/ContactController.js
@@ -1,5 +1,6 @@
 // const { response } = require('express');
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -12,6 +13,10 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid ccontact id' });
+    }
 
     const contact = await ContactsRepository.findById(id);
 
@@ -27,16 +32,25 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' });
+    }
+
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExistis = await ContactsRepository.findByEmail(email);
-    if (contactExistis) {
-      return response.status(400).json({ error: 'This email is already in use' });
+    if (email) {
+      const contactExistis = await ContactsRepository.findByEmail(email);
+      if (contactExistis) {
+        return response.status(400).json({ error: 'This email is already in use' });
+      }
     }
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -48,23 +62,34 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExistis = await ContactsRepository.findById(id);
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid ccontact id' });
+    }
 
-    if (!contactExistis) {
-      return response.status(404).json({ error: 'Contact not found' });
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
+    const contactExistis = await ContactsRepository.findById(id);
+    if (!contactExistis) {
+      return response.status(404).json({ error: 'Contact not found' });
+    }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: ' This email is already in use' });
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: ' This email is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -72,6 +97,10 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid ccontact id' });
+    }
 
     await ContactsRepository.delete(id);
     response.sendStatus(204);
